requirements.txt
streamlit
torch
torchvision
numpy
Pillow
huggingface-hub
facenet-pytorch
scikit-learn
opencv-python-headless